app/requirements.txt
SQLAlchemy~=2.0.28
alembic~=1.13.1
fastapi~=0.110.0
asyncpg
uvicorn
gunicorn
pydantic_settings
scikit-surprise~=1.1.4